lot of changes in the api

+++ server/SafeShare/SafeShare.API/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Identity.Data;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity.Data;
 using Microsoft.AspNetCore.Mvc;
 using SafeShare.API.Models;
 using SafeShare.CORE.Entities;
@@ -9,6 +10,8 @@ namespace SafeShare.API.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
+
     public class UsersController : ControllerBase
     {
         private readonly IUserService _userService;
@@ -17,34 +20,16 @@ namespace SafeShare.API.Controllers
             _userService = userService;
         }
 
-        [HttpPost("register")]
-        public async Task<IActionResult> RegisterAsync([FromBody] UserPostModel user)
-        {
-            var userSend = new User()
-            { Email = user.Email, PasswordHash = user.PasswordHash, IsAdmin = user.IsAdmin, Username = user.Username };
-
-            var result = await _userService.RegisterAsync(userSend);
-            if (result != null)  // אם ההרשמה הצליחה
-                return CreatedAtAction(nameof(GetUserByIdAsync), new { userId = result.UserId }, result);
-            return BadRequest(new { message = "Registration failed", errors = "Some error message or validation issue." });
-        }
-
-        // פונקציה להתחברות
-        [HttpPost("login")]
-        public async Task<IActionResult> LoginAsync([FromBody] LoginRequest loginRequest)
-        {
-            
-            var user = new User() { Email = loginRequest.Email, PasswordHash = loginRequest.Password };
-            var result = await _userService.LoginAsync(user);
-            if (result == null)  // אם ההתחברות נכשלה
-                return Unauthorized(new { message = "Invalid credentials" });
-            return Ok(result); // מחזירים את המידע על המשתמש
-        }
+  
 
         // עדכון משתמש
         [HttpPut("{userId}")]
         public async Task<IActionResult> UpdateUserAsync(int userId, [FromBody] UserPostModel user)
         {
+            var isAdminClaim = User.FindFirst("isAdmin")?.Value;
+            var idClaim = User.FindFirst("id")?.Value;
+            if (isAdminClaim == "false" || userId.ToString() != idClaim)
+                return Unauthorized();
             var userSend = new User() { Email = user.Email, PasswordHash = user.PasswordHash, IsAdmin = user.IsAdmin, Username = user.Username };
             var result = await _userService.UpdateUserAsync(userId, userSend);
             if (result != null)  // אם העדכון הצליח
@@ -56,6 +41,9 @@ namespace SafeShare.API.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAllUsersAsync()
         {
+            var isAdminClaim = User.FindFirst("isAdmin")?.Value;
+            if(isAdminClaim == "false")
+                return Unauthorized();
             var users = await _userService.GetAllUsersAsync();
             if (users != null && users.Any())
                 return Ok(users);
@@ -66,6 +54,10 @@ namespace SafeShare.API.Controllers
         [HttpGet("{userId}")]
         public async Task<IActionResult> GetUserByIdAsync(int userId)
         {
+            var isAdminClaim = User.FindFirst("isAdmin")?.Value;
+            var idClaim = User.FindFirst("id")?.Value;
+            if (isAdminClaim == "false"||userId.ToString()!=idClaim)
+                return Unauthorized();
             var user = await _userService.GetUserByIdAsync(userId);
             if (user != null)
                 return Ok(user);
@@ -76,6 +68,10 @@ namespace SafeShare.API.Controllers
         [HttpDelete("{userId}")]
         public async Task<IActionResult> DeleteUserAsync(int userId)
         {
+            var isAdminClaim = User.FindFirst("isAdmin")?.Value;
+            var idClaim = User.FindFirst("id")?.Value;
+            if (isAdminClaim == "false" || userId.ToString() != idClaim)
+                return Unauthorized();
             var result = await _userService.DeleteUserAsync(userId);
             if (result)  // אם המחיקה הצליחה
                 return Ok(new { message = "User deleted successfully" });
